feat: post process report (#446)
## Describe the issue:
Sometimes either users or the bot submit reports that are:
1. incomplete 
2. impossible
We need to validate those, to make sure that we only have valid and as
complete as possible reports in the database.

## Explain how you solved the issue:
I defined a pipe that calls a couple of cross validation function on the
report that was being sent by the user.

## Additional notes or considerations:
A lot of this is coming from the extensive testing. If you have any
ideas what we should also test in this context please let me know.
I also tried to write this as functional as possible so that we can
reduce complexity.

### Important!!!
`guessStation` is not scaleable. It will fetch all of the reports for
the line into memory and choose the most common in memory. This means as
we get more reports the memory (and also CPU) usage of posting an
inspector will increase.
There are a couple of ways how we can solve this each with their own
trade off and I would like some opinions on what we should do:
1. <del>Move the entire calculation to the DB. We can craft one drizzle
query that will perform this all on the db side. The benefit of this is
that this is as performant as it can possibly get since the brightest
minds in computer science are working to optimize postgres and no matter
how hard we try we will not be faster than Postgres. The downside of
this is that the query would be quite complex, meaning difficult to
write and even more difficult to read. </del>
2. <del>Create DB aggregates. We can create some aggregates that store
the most common station for each line for each day of the week and each
hour. We can then update this with a trigger every 24h at midnight or
something like this. This would mean we have a O(1) lookup time but we
also increase complexity of the db schema significantly. The benefit of
this is that we can reuse the same logic for historic data, which means
that the `POST /reports/` and the `GET /reports/` both benefit from
this.</del>
3. Limit results: We can simply return the results of the last 30 or 90
days, with which we would have a very simple solution that is also very
fast. The issue with this is that we loose significant accuracy
especially if we would deploy in a city with a less active community.
4.<del> Do nothing. We currently have around 80k reports, even if all of
them are loaded into memory it would not be that big of deal especially
since we are not getting that many reports concurrently the server can
probably handle it for now. So we could just note this down and solve
the issue once it becomes an actual issue. </del>

Also depending on the decision I can make `guessStation` be pure.